Working with Exponential Filter
Former-commit-id: 46cd784c3dc0a9392455f6f25940809821929b7f [formerly e0bb957a76467e5b0b33132118c42f7bf9041af2]
Former-commit-id: 804307cce5e0e75035a7dab4c68e97f6154dda11
Former-commit-id: 7d355e553a31cfb1a7335766dd03e1fe9e362434

+++ src/library/natrium/solver/SolverConfiguration.h
@@ -1022,7 +1022,7 @@ public:
 		return s;
 	}
 
-	void setExponentialFilterNc(long int Nc) {
+	void setExponentialFilterNc(double Nc) {
 		enter_subsection("Filtering");
 		enter_subsection("Filter parameters");
 		try {
